Restructure the library, introduce support for peer metadata, GC for lost peers, lost peer notifications, shutdown support, fix error shadowing.
